united tests for changePassword form in one file
