+++ src/components/changePassword/changePassword.test.js
@@ -1,8 +1,26 @@
 import React from 'react';
 import ChangePassword from './index';
 import userEvent from '@testing-library/user-event';
-import { render, wait } from '@testing-library/react';
+import { render, wait, fireEvent } from '@testing-library/react';
 import { BrowserRouter as Router } from 'react-router-dom';
+import { AuthProvider } from '../../services/tokenValidator';
+
+jest.mock('../../services/useCaptcha.tsx', () => ({
+  _esModule: true,
+  useCaptcha: jest.fn().mockImplementation(() => ({
+    getCaptchaToken: jest
+      .fn()
+      .mockImplementation(() => Promise.resolve('QWerty1@')),
+  })),
+}));
+
+const mockJsonPromise = Promise.resolve('received data');
+const mockFetchPromise = Promise.resolve({
+  json: () => mockJsonPromise,
+  status: 200,
+});
+
+global.fetch = jest.fn().mockImplementation(() => mockFetchPromise);
 
 describe('ChangePassword', () => {
   test('render component correctly', () => {
@@ -32,7 +50,7 @@ describe('ChangePassword', () => {
       getByPlaceholderText('Підтвердіть новий пароль'),
       'Qwerty-1'
     );
-    
+
     await wait(() => {
       expect(getByPlaceholderText('Новий пароль')).toHaveValue('Qwerty-1');
       expect(getByPlaceholderText('Підтвердіть новий пароль')).toHaveValue(
@@ -59,5 +77,21 @@ describe('ChangePassword', () => {
     await wait(() => {
       expect(queryByText('Паролі мають співпадати')).toBeInTheDocument();
     });
+  });
+  test('shows a notification of a successful password change', async () => {
+    const { getByPlaceholderText, getByRole } = render(
+      <AuthProvider>
+        <ChangePassword />
+      </AuthProvider>
+    );
+
+    userEvent.type(getByPlaceholderText('Старий пароль'), 'Qwerty1@');
+    userEvent.type(getByPlaceholderText('Новий пароль'), 'Qwerty1@2');
+    userEvent.type(
+      getByPlaceholderText('Підтвердіть новий пароль'),
+      'Qwerty1@2'
+    );
+    fireEvent.click(getByRole('button'));
+    await wait(() => expect(global.fetch).toHaveBeenCalled());
   });
 });
